refactor:merge route params
