--- src/screens/Projects/types.ts
@@ -1,5 +1,0 @@
-export interface RouteParams {
-  route: {
-    params: Record<string, unknown> | undefined;
-  };
-}
