--- src/screens/Splash/types.ts
@@ -1,7 +1,0 @@
-type RouteParams = string | {name: string; params?: Record<string, unknown>};
-
-export interface NavigationProps {
-  navigation: {
-    navigate: (routeParams: RouteParams) => {};
-  };
-}
